requirements.txt
fastapi
uvicorn
sqlalchemy
sqlmodel
alembic
psycopg2-binary
presidio-analyzer
presidio-anonymizer
streamlit
streamlit-aggrid
plotly
pandas
openpyxl
python-docx
pymupdf
apscheduler
celery
redis
sqlglot
requests
pyjwt
passlib[bcrypt]
structlog
python-multipart
# New additions based on feedback
scikit-learn
datasketch
pyhive
boto3
pytesseract
great_expectations
pillow
